requirements.txt
pi==0.1.2
Pillow==9.5.0
python-dotenv==1.0.0
